utils: Move Git index lock waiting
utils.git now gets optional-lock environment setup from
utils.git_environment while still owning Git command execution and
index-lock waiting.

Index-lock path discovery and polling are a separate command-scheduling
concern from running Git subprocesses. Keeping that wait loop inside
utils.git makes the module harder to scan as more Git helpers accumulate.

This commit addresses that by moving index-lock path discovery and waiting
into utils.git_index_lock and routing run_git_command through that module.

The following tests commit will move Git utility coverage onto the new
index-lock module.
