+++ src/git_stage_batch/utils/git.py
@@ -14,6 +14,7 @@ from pathlib import Path
 
 from ..exceptions import exit_with_error
 from ..i18n import _
+from . import git_index_lock
 from .command import ExitEvent, OutputEvent, run_command, stream_command
 from .git_environment import git_environment_with_optional_locks_disabled
 from .text import bytes_to_lines
@@ -21,8 +22,6 @@ from .text import bytes_to_lines
 
 _GIT_REPOSITORY_ROOT_CACHE: dict[Path, Path] = {}
 _GIT_DIRECTORY_CACHE: dict[Path, Path] = {}
-_INDEX_LOCK_WAIT_SECONDS = 20.0
-_INDEX_LOCK_POLL_SECONDS = 0.05
 
 
 @dataclass(frozen=True)
@@ -44,57 +43,6 @@ class GitIndexEntryUpdate:
     force_remove: bool = False
 
 
-def _custom_index_lock_path(
-    *,
-    env: dict[str, str] | None,
-    cwd: str | None,
-) -> Path | None:
-    git_env = os.environ.copy() if env is None else dict(env)
-    index_file = git_env.get("GIT_INDEX_FILE")
-    if not index_file:
-        return None
-
-    index_path = Path(index_file)
-    if not index_path.is_absolute():
-        index_path = (Path.cwd() if cwd is None else Path(cwd)) / index_path
-    return Path(f"{index_path}.lock")
-
-
-def _git_index_lock_path(*, cwd: str | None, env: dict[str, str] | None) -> Path:
-    custom_index_lock_path = _custom_index_lock_path(env=env, cwd=cwd)
-    if custom_index_lock_path is not None:
-        return custom_index_lock_path
-
-    result = run_command(
-        ["git", "rev-parse", "--absolute-git-dir"],
-        check=True,
-        cwd=cwd,
-        env=git_environment_with_optional_locks_disabled(env),
-    )
-    return Path(result.stdout.strip()) / "index.lock"
-
-
-def wait_for_git_index_lock(
-    *,
-    cwd: str | None = None,
-    env: dict[str, str] | None = None,
-    timeout_seconds: float = _INDEX_LOCK_WAIT_SECONDS,
-    poll_seconds: float = _INDEX_LOCK_POLL_SECONDS,
-) -> None:
-    """Wait briefly for a pre-existing Git index lock to disappear."""
-    try:
-        index_lock_path = _git_index_lock_path(cwd=cwd, env=env)
-    except subprocess.CalledProcessError:
-        return
-
-    deadline = time.monotonic() + timeout_seconds
-    while index_lock_path.exists():
-        remaining_seconds = deadline - time.monotonic()
-        if remaining_seconds <= 0:
-            return
-        time.sleep(min(poll_seconds, remaining_seconds))
-
-
 def _prepare_git_command_environment(
     *,
     requires_index_lock: bool,
@@ -102,7 +50,7 @@ def _prepare_git_command_environment(
     env: dict[str, str] | None,
 ) -> dict[str, str] | None:
     if requires_index_lock:
-        wait_for_git_index_lock(cwd=cwd, env=env)
+        git_index_lock.wait_for_git_index_lock(cwd=cwd, env=env)
     return _git_command_environment(requires_index_lock=requires_index_lock, env=env)
 
 
@@ -345,13 +293,13 @@ def run_git_command(
         else stdin_chunks
     )
     retry_deadline = (
-        time.monotonic() + _INDEX_LOCK_WAIT_SECONDS
+        time.monotonic() + git_index_lock.DEFAULT_INDEX_LOCK_WAIT_SECONDS
         if requires_index_lock
         else None
     )
 
     if retry_deadline is not None:
-        wait_for_git_index_lock(
+        git_index_lock.wait_for_git_index_lock(
             cwd=cwd,
             env=env,
             timeout_seconds=_remaining_index_lock_wait_seconds(retry_deadline),
@@ -381,7 +329,7 @@ def run_git_command(
         remaining_seconds = _remaining_index_lock_wait_seconds(retry_deadline)
         if remaining_seconds <= 0:
             break
-        wait_for_git_index_lock(
+        git_index_lock.wait_for_git_index_lock(
             cwd=cwd,
             env=env,
             timeout_seconds=remaining_seconds,
